rename display utils
